Move indices & edges for each move in move not map
Move notation maps now holds the indices and edges for each move. It is
not yet working as intended. For it to work the following must be done:
Update each move func to use the indices & edges from the move notation
map. Swap the indices and edges along with moves on update move notation
function.
other moves